pyproject.toml
[tool.poetry]
name = "writestackai"
version = "0.1.0"
description = ""
authors = ["lil-sussy <naykeysnet@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
django-rest-framework = "^0.1.0"
chromadb = "^0.5.0"
logger = "^1.4"
requests = "^2.32.2"
colorlog = "^6.8.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
